fix : 버그 잡는 기계

+++ Godslayer-New-Age/Godslayer-New-Age/Scenes/MapleScene.cs
@@ -294,12 +294,15 @@ internal class MapleScene : IScene
         PrintUtil.CreateBox();
         while (true)
         {
-            if (int.TryParse(Console.ReadLine(), out input3)&&(input3 >= 0 && input3 <= 2))
+            if (int.TryParse(Console.ReadLine(), out input3) && (input3 >= 0 && input3 <= 2))
             {
-                break;
+                return input3;
+            }
+            else
+            {
+                return 100;
             }
         }
-        return input3;
     }
     public int TargetText2(int Scene)
     {
@@ -314,10 +317,13 @@ internal class MapleScene : IScene
         {
             if (int.TryParse(Console.ReadLine(), out input3) && (input3 >= 0 && input3 <= 1))
             {
-                break;
+                return input3;
+            }
+            else
+            {
+                return 100;
             }
         }
-        return input3;
     }
     public int PlayerSkillText(int Scene)
     {
@@ -332,10 +338,13 @@ internal class MapleScene : IScene
             int.TryParse(Console.ReadLine(), out input2);
             if (input2 >= 1 && input2 <= 4)
             {
-                break;
+                return input2 - 1;
+            }
+            else
+            {
+                return 100;
             }
         }
-        return input2 - 1;
     }
     public void PressAnyKey(int Scene)
     {
@@ -417,7 +426,15 @@ internal class MapleScene : IScene
                 {
                     box1Text[2].Add(" ");
                     int skillnum = PlayerSkillText(2);
+                    if (skillnum == 100)
+                    {
+                        return GameState.Retry;
+                    }
                     int targetnum = TargetText(2);
+                    if (targetnum == 100)
+                    {
+                        return GameState.Retry;
+                    }
                     if (box1Text[2].Count > 18)
                     {
                         box1Text[2].RemoveRange(2, box1Text.Count - 2);
@@ -470,7 +487,15 @@ internal class MapleScene : IScene
                 while (Player.Instance.HP > 0 && monsters[4].HP > 0)
                 {
                     int skillnum = PlayerSkillText(3);
+                    if (skillnum == 100)
+                    {
+                        return GameState.Retry;
+                    }
                     int targetnum = TargetText2(3);
+                    if (targetnum == 100)
+                    {
+                        return GameState.Retry;
+                    }
                     box1Text[3].RemoveRange(19, box1Text[3].Count - 19);
                     StartBattle(monsters[4], skillnum, targetnum);
                 }
@@ -511,7 +536,15 @@ internal class MapleScene : IScene
                 while (Player.Instance.HP > 0 && monsters[5].HP > 0)
                 {
                     int skillnum = PlayerSkillText(4);
+                    if (skillnum == 100)
+                    {
+                        return GameState.Retry;
+                    }
                     int targetnum = TargetText2(4);
+                    if (targetnum == 100)
+                    {
+                        return GameState.Retry;
+                    }
                     box1Text[4].RemoveRange(19, box1Text[4].Count - 19);
                     StartBattle(monsters[5], skillnum, targetnum);
                 }
